Adding the last fixes on the pipes and redirections handling;
