+++ exec/exec.h
@@ -28,21 +28,14 @@ bool	is_builtin(t_ast *ast);
 bool	is_pipe_next(t_ast *ast);
 int		setup_redirections(t_ast *ast);
 bool	is_pipe(t_ast *ast);
-int		init_gates(t_ast *ast);
-int		setup_fds(t_ast *ast, int pipe_in, int pipe_out);
 char	*get_full_name(char *name);
 bool	is_logical_connector(t_ast *ast);
 int	count_nodes(t_ast *ast);
 
-/* Gates (Pipelines + redirections) */
-int		setup_gates(t_ast *ast, int node_count);
-int		restore_gates(void);
-int		close_gates(void);
 
 /* Pids and processes management */
 int		pid_wait_all(void);
 int		pid_push(pid_t pid);
-int		pid_wait_last(void);
 void	pid_init(void);
 
 /* Executer functions */
@@ -52,17 +45,12 @@ int		exec_executable(t_ast *ast);
 int		exec(t_ast *ast);
 t_ast	*exec_connector(t_ast *ast, int *node_count);
 
-/* Pipes Getters/setters */
-int		pipe_init(void);
 
 /* Static declarations settings */
 int		pipe_in_act(int action, int hold);
 int		pipe_out_act(int action, int hold);
-int		current_pipe_act(int action, int hold);
-int		pipe_count_act(int action, int hold);
-int		**pipes_act(int action);
 
-/* New implementation of the pipes */
+/* implementation of the pipes --> pipe.c */
 int	**init_pipes(int *number_of_nodes, t_ast *ast);
 int	setup_pipes(int count, int **pipes, int number_of_nodes);
 int	close_all_pipes(int **pipes, int count);
