+++ exec/getters.c
@@ -21,13 +21,3 @@ int	get_pipe_out(void)
 {
 	return (pipe_out_act(GET, 0));
 }
-
-int	get_pipe_count(void)
-{
-	return (pipe_count_act(GET, 0));
-}
-
-int	get_current_pipe(void)
-{
-	return (current_pipe_act(GET, 0));
-}
