+++ exec/statics.c
@@ -34,62 +34,3 @@ int	pipe_out_act(int action, int hold)
 		pipe_out = hold;
 	return (pipe_out);
 }
-
-int	current_pipe_act(int action, int hold)
-{
-	static int	current_pipe;
-
-	if (action == GET)
-		return (current_pipe);
-	if (action == SET)
-		current_pipe = hold;
-	return (current_pipe);
-}
-
-int	pipe_count_act(int action, int hold)
-{
-	static int	pipe_count;
-
-	if (action == GET)
-		return (pipe_count);
-	if (action == SET)
-		pipe_count = hold;
-	return (pipe_count);
-}
-
-int	**pipes_act(int action)
-{
-	static int	**pipes;
-	int			i;
-
-	if (action == SET)
-	{
-		pipe_count_act(SET, 0);
-		current_pipe_act(SET, 0);
-		pipes = NULL;
-		return (NULL);
-	}
-	if (pipes == NULL)
-	{
-		/* printf("[ALLOCATING PIPES WITH SIZE %d]\n", MAX_PIPES_COUNT); */
-		pipes = ft_malloc(sizeof(int *), MAX_PIPES_COUNT);
-		if (!pipes)
-			exit(0);
-		i = 0;
-		while (i < MAX_PIPES_COUNT)
-		{
-			pipes[i] = ft_malloc(sizeof(int), PIPE_SIZE);
-			if (!pipes[i])
-				exit(0);
-			i++;
-		}
-		i = 0;
-		while (i < MAX_PIPES_COUNT)
-		{
-			pipes[i][0] = -1;
-			pipes[i][1] = -1;
-			i++;
-		}
-	}
-	return (pipes);
-}
